--- src/main/java/ListCommand.java
@@ -1,6 +1,0 @@
-public class ListCommand extends Command {
-    @Override
-    public void execute(Storage storage, TaskList taskList) {
-        taskList.printTaskList();
-    }
-}
