--- src/main/java/Tool.java
@@ -1,9 +1,0 @@
-public class Tool {
-
-
-
-
-
-
-
-}
