+++ src/main/java/dino/exception/DinoException.java
@@ -1,0 +1,31 @@
+package dino.exception;
+
+public class DinoException extends Exception {
+    public DinoException(String message) {
+        super(message);
+    }
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
