+++ src/main/java/dino/exception/EmptyListException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class EmptyListException extends DinoException {
+    public EmptyListException() {
+        super("😐 You don't have any task in your list~");
+    }
+}
