+++ src/main/java/dino/exception/EmptyTaskDescriptionException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class EmptyTaskDescriptionException extends DinoException {
+    public EmptyTaskDescriptionException() {
+        super("😶 The description of a task cannot be empty.");
+    }
+}
