+++ src/main/java/dino/exception/IndexNotSpecifiedException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class IndexNotSpecifiedException extends DinoException {
+    public IndexNotSpecifiedException() {
+        super("😕 Please specify the index of the task!");
+    }
+}
