+++ src/main/java/dino/exception/InvalidFormatException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class InvalidFormatException extends DinoException {
+    public InvalidFormatException(String action, String format) {
+        super("😕 Please "  + action + " is entered in the format: " + format);
+    }
+}
