+++ src/main/java/dino/exception/InvalidIndexException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class InvalidIndexException extends DinoException {
+    public InvalidIndexException() {
+        super("😮 Please enter a valid task index!");
+    }
+}
