+++ src/main/java/dino/exception/InvalidInputException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class InvalidInputException extends DinoException {
+    public InvalidInputException() {
+        super("🤔 Hmm... sorry I really don't understand what you mean.");
+    }
+}
