+++ src/main/java/dino/exception/TaskAlreadyDoneException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class TaskAlreadyDoneException extends DinoException {
+    public TaskAlreadyDoneException() {
+        super("🤨 Hey you have already done this task!");
+    }
+}
