+++ src/main/java/dino/exception/TimeNotSpecifiedException.java
@@ -1,0 +1,7 @@
+package dino.exception;
+
+public class TimeNotSpecifiedException extends DinoException {
+    public TimeNotSpecifiedException(String type) {
+        super("😛 Please specify the time of your " + type + "!");
+    }
+}
