+++ src/main/java/dino/task/Event.java
@@ -1,6 +1,9 @@
+package dino.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+import dino.exception.*;
 
 public class Event extends Task{
