+++ src/main/java/dino/task/Task.java
@@ -1,3 +1,5 @@
+package dino.task;
+
 public class Task {
 
     private String description;
