+++ src/main/java/dino/task/ToDo.java
@@ -1,3 +1,5 @@
+package dino.task;
+
 public class ToDo extends Task {
     public ToDo(String description) {
         super(description);
