+++ src/main/java/dino/util/Ui.java
@@ -1,4 +1,7 @@
+package dino.util;
+
 import java.util.Scanner;
+import dino.task.TaskList;
 
 public class Ui {
